clh: api update
update api client with latest cloud hypervisor.

+++ virtcontainers/pkg/cloud-hypervisor/client/model_fs_config.go
@@ -14,5 +14,5 @@ type FsConfig struct {
 	Sock string `json:"sock"`
 	NumQueues int32 `json:"num_queues"`
 	QueueSize int32 `json:"queue_size"`
-	CacheSize int32 `json:"cache_size,omitempty"`
+	CacheSize int64 `json:"cache_size,omitempty"`
 }
